+++ headers/entities/geometry/Geometry.hpp
@@ -16,7 +16,7 @@ template<typename type>
 struct ThreeNum_set { 
 	//type for 3 num sets
 	type X, Y, Z;
-	ThreeNum_set() : X(type(0.0)), Y(type(0.0)), Z(type(0.0)) {}; //general constructor
+	ThreeNum_set() : X(type(0)), Y(type(0)), Z(type(0)) {}; //general constructor
 	ThreeNum_set(type x_val, type y_val, type z_val) : X(type(x_val)), Y(type(y_val)), Z(type(z_val)) {}; //specific constructor
 };
 
@@ -24,7 +24,7 @@ template<typename type>
 struct TwoNum_set { 
 	//type for 2 num sets
 	type X, Y;
-	TwoNum_set() : X(type(0.0)), Y(type(0.0)) {};
+	TwoNum_set() : X(type(0)), Y(type(0)) {};
 	TwoNum_set(type x_val, type y_val) : X(type(x_val)), Y(type(y_val)) {};
 };
 
@@ -41,6 +41,6 @@ struct TwoNum_set {
 
 #pragma endregion
 
-bool compareThreeNum_set (ThreeNum_set <float> firstSet, ThreeNum_set<float> secondSet);
+bool compareThreeNum_set (ThreeNum_set <int> firstSet, ThreeNum_set<int> secondSet);
 
 #endif
